concurrent_queue implemented

+++ merge.h
@@ -1,3 +1,3 @@
 #include <unordered_map>
 
-void merge(std::unordered_map<std::string, int> dict1, std::unordered_map<std::string, int> dict2);
+void merge(std::unordered_map<std::string, size_t> dict1, std::unordered_map<std::string, size_t> dict2);
